refactored components of Rule Collection into separate vue files
