+++ client/galaxy/scripts/components/RuleBuilder/IdentifierDisplay.vue
@@ -1,0 +1,50 @@
+<template>
+    <li class="rule">
+        <span v-b-tooltip.hover :title="help">Set {{ columnsLabel }} as {{ typeDisplay }}</span>
+        <span v-b-tooltip.hover :title="titleEdit" class="fa fa-edit" @click="edit"></span>
+        <span v-b-tooltip.hover :title="titleRemove" class="fa fa-times" @click="remove"></span>
+    </li>
+</template>
+
+<script>
+export default{
+    props: {
+        type: {
+            type: String,
+            required: true,
+        },
+        columns: {
+            required: true,
+        },
+        colHeaders: {
+            type: Array,
+            required: true,
+        },
+    },
+    methods: {
+        remove() {
+            this.$emit("remove");
+        },
+        edit() {
+            this.$emit("edit");
+        },
+    },
+    computed: {
+        typeDisplay() {
+            return MAPPING_TARGETS[this.type].label;
+        },
+        help() {
+            return MAPPING_TARGETS[this.type].help || "";
+        },
+        titleEdit() {
+            return _l("Edit column definition");
+        },
+        titleRemove() {
+            return _l("Remove this column definition");
+        },
+        columnsLabel() {
+            return RuleDefs.columnDisplay(this.columns, this.colHeaders);
+        },
+    }
+}
+</script>
